max priority->min, and admin function
lower priority is now higher priority, so change max_priority
to min.
make it read this value from a file and add an admin function
for changing it.
this lets an admin specify a mimimum priority an intersection
must have to be available for signup so that the highest
(er, lowest) priority intersections can be assigned first.